Update the dockerfile and build instructions
The dockerfile needed updating since the base OS version was no-longer
being maintained, and packages for the version unavailable to download.

This commit updates the Dockerfile to use Ubuntu 22.04 as the base image.
I also removed dependency on the horgix/moderncv image as it is no-longer
being maintained.

Since the dockerfile can be used with podman, update the build instructions
in the README to podman commands. Podman is preferred as it doesn't require
a daemon to run and can be run by a non-root user.

Signed-off-by: Jason Rogena <jrogena@fb.com>